==> 0_extrair_dados.ipynb <==
# Módulos

import numpy as np
import pandas as pd

# __Execução__

# Definir caminhos de dados
path_enem_19 = 'dados/microdados/microdados_enem_2019/DADOS/MICRODADOS_ENEM_2019.csv' 
path_enem_20 = 'dados/microdados/microdados_enem_2020/DADOS/MICRODADOS_ENEM_2020.csv' 
path_enem_21 = 'dados/microdados/microdados_enem_2021/DADOS/MICRODADOS_ENEM_2021.csv' 

##### Puxar + Filtrar

# Extrair base completa
df_enem = pd.read_csv(path_enem_19, low_memory=False, sep=';', encoding='latin-1')

# Definir condição de seleção de alunos para análise
condicao = df_enem['SG_UF_PROVA'] == 'BA'
# Extrair
df_e19_ba = df_enem[condicao].copy()

# Extrair base completa
df_enem = pd.read_csv(path_enem_20, low_memory=False, sep=';', encoding='latin-1')

# Definir condição de seleção de alunos para análise
condicao = df_enem['SG_UF_PROVA'] == 'BA'

# Extrair
df_e20_ba = df_enem[condicao].copy()

# Extrair base completa
df_enem = pd.read_csv(path_enem_21, low_memory=False, sep=';', encoding='latin-1')

# Definir condição de seleção de alunos para análise
condicao = df_enem['SG_UF_PROVA'] == 'BA'

# Extrair
df_e21_ba = df_enem[condicao].copy()

#### Salvar em memória

# Salvar localmente
df_e19_ba.to_csv('dados/extracao/enem_2019_sg_uf_prova_ba.csv', index=False)
df_e20_ba.to_csv('dados/extracao/enem_2020_sg_uf_prova_ba.csv', index=False)
df_e21_ba.to_csv('dados/extracao/enem_2021_sg_uf_prova_ba.csv', index=False)
