==> 3_analise_dados_pratica.ipynb <==
# Módulos

# Padrão
import numpy as np
import pandas as pd

# Visualização
import seaborn as sns
import matplotlib.pyplot as plt

# Análise estatística
from scipy import stats

# Extração de dados

# Definir caminhos de dados
path_enem = 'dados/enem_analise_completa.csv' 
path_enem_presenca = 'dados/enem_analise_presenca_completa.csv'  

# Extrair base completa
df_enem = pd.read_csv(path_enem, low_memory=False, sep=',', encoding='utf-8')
df_enem_nota = pd.read_csv(path_enem_presenca, low_memory=False, sep=',', encoding='utf-8')

# Análises
